AlphaConformers

Documentation for AlphaConformers.

AlphaConformers._read_usalign_output_table

AlphaConformers.create_pdb_folder

AlphaConformers.foldseek_search

AlphaConformers.read_foldseek_search_results

AlphaConformers.usalign

AlphaConformers._read_usalign_output_table — Method

Parse the table output format of USalign (-outfmt 2) into a DataFrame

AlphaConformers.create_pdb_folder — Method

create_pdb_folder(targets::Union{Set, DataFrames.DataFrame, AbstractArray}, 
                  folder_path::String; 
                  pdb_db::String = get(ENV, "PDB_DB", ""), 
                  extension::String = ".pdb")

Creates a local folder at folder_path and populates it with the PDB files specified by the targets. The targets argument can be a Set, an AbstractArray, or a DataFrame with a target column, such as the results loaded from a Foldseek search using the read_foldseek_search_results function. Each target represents a PDB code, possibly followed by a chain code. If a chain code is specified, only that chain is saved in the new folder. If no chain code is specified, all chains are saved. The expected format of the targets depends on the number of chains:

1 chain: "pdb"

2 chains: "pdb_chain"

For example: "1EX7.pdb" or "4F4J.pdb_B". Note that the presence of the "pdb" extension is not mandatory and depends on the database used when reading Foldseek results.

The PDB files are either fetched from a local database path specified by pdb_db or downloaded from the web if pdb_db is an empty string. Note that you can set the pdb_db keyword argument when calling the function, or alternatively, set the PDB_DB environment variable. When retrieving the PDB files from a local folder, the optional extension argument specifies the file extension of the PDB files in that folder (default is ".pdb").

Warning

If the specified folder_path already exists, it will be deleted and a new folder will be created.

The function returns the absolute path to the created folder.

AlphaConformers.foldseek_search — Method

foldseek_search(pdb_file::AbstractString; db_path::String = get(ENV, "FOLDSEEK_DB_PATH", ""))

Searches a given protein structure in the Foldseek database.

The pdb_file is the path to the protein structure file in PDB format. The db_path is the path to the Foldseek database, which defaults to the FOLDSEEK_DB_PATH environment variable. If the FOLDSEEK_DB_PATH environment variable is not set and db_path is not given, the function will throw an error.

The function executes a Foldseek easy-search, saving the results in an output file with the same base name as the PDB file but with _results.m8 appended to it. The function operates in a temporary directory that is automatically cleaned up afterwards.

The function returns the path to the Folseek easy-search output file.

AlphaConformers.read_foldseek_search_results — Method

read_foldseek_search_results(file::AbstractString)

Reads the Foldseek easy-search output file (m8) and returns a DataFrame with the results and proper column names.

AlphaConformers.usalign — Method

usalign(pdb_file_a, pdb_file_b)
usalign(pdb_file_a, pdb_folder, pdb_list)

Performs structural alignment of protein structures using the USalign algorithm. The alignment is performed using the -fast option, which is slightly less accurate but faster than the default alignment method. The -TMcut 0.5 option is used to skip alignments unlikely to reach a TM-score of 0.5.

Function signatures

usalign(pdb_file_a, pdb_file_b) calculates the structural alignment between two protein structures. It takes the path to the two protein structures in PDB format as inputs.

usalign(pdb_file_a, pdb_folder, pdb_list) calculates the structural alignment between one protein structure and multiple other protein structures. It takes the path to the protein structure in PDB format to be aligned (pdb_file_a), the path to the directory containing the protein structures to align with (pdb_folder), and a list of protein structures (in PDB format) contained in pdb_folder (pdb_list) as inputs. You can use the create_pdb_folder function to create the pdb_folder using a list of target protein structures.

In both cases, the function asserts the existence of the input file(s) before proceeding with the alignment and returns a DataFrame with the results of the alignment.

USalign parameters

This function uses the USalign tool, interfaced via the USalign_jll package. The actual alignment is performed by running the tool in a subprocess, and the output is parsed from the tool's tabular output format (-outfmt 2). This function uses the following default parameters when running USalign:

-mol prot: Align proteins only

-TMcut 0.5: Skip alignment if TM-score is unlikely to reach 0.5

-fast: Use a faster but slightly less accurate alignment method

-outfmt 2: Use the tabular format for compact output that can be parsed with CSV.jl

Those parameters are also keyword arguments to the usalign function. You can use those keyword arguments to change this default behavior. To avoid using the TMscore cutoff, you should set TMcut to nothing. To use the slower but more accurate alignment method, you should set fast to false.

Also, you can pass a Cmd object to the `` keyword argument to add or modify other USalign arguments. Please note that you do not need to give the structures to align, as they are automatically passed to USalign.