AlphaConformers

Documentation for AlphaConformers.

AlphaConformers.StructuralClustering

AlphaConformers._create_empty_folder

AlphaConformers._get_abspath

AlphaConformers._read_usalign_output_table

AlphaConformers.align_sequences

AlphaConformers.clean_msa

AlphaConformers.create_msa_and_templates

AlphaConformers.create_pdb_folder

AlphaConformers.create_pdb_lists

AlphaConformers.foldseek_search

AlphaConformers.get_clustered_pdbs

AlphaConformers.get_pdb_codes

AlphaConformers.get_residues_and_sequence

AlphaConformers.get_uniprot_acc

AlphaConformers.get_uniprot_mapping

AlphaConformers.read_foldseek_search_results

AlphaConformers.run_alphafold

AlphaConformers.save_sequences

AlphaConformers.usalign

AlphaConformers.StructuralClustering — Type

StructuralClustering

Struct to store the results of the structural clustering. It contains the absolute paths to the PDB files (pdbs), the cluster number assigned to each file (clusters), the size of each cluster (cluster_sizes), and the total number of clusters (nclusters). This struct is a subtype of Clustering.ClusteringResult and implements its interface by defining the nclusters, counts, and assignments methods.

AlphaConformers._create_empty_folder — Method

_create_empty_folder(path)

Helper function to create an empty folder at the given path. If the folder already exists, all its contents will be deleted.

AlphaConformers._get_abspath — Method

_get_abspath(list_item, pdb_folder)

Helper function to get the absolute path of a file. This is useful as the pdb_list for usalign are relative paths to a subset of the files in pdb_folder. However, the first argument of usalign must be an absolute path.

AlphaConformers._read_usalign_output_table — Method

Parse the table output format of USalign (-outfmt 2) into a DataFrame

AlphaConformers.align_sequences — Method

align_sequences(sequences::String)

This function uses MAFFT (thanks to the MAFFT_jll package) to align the sequences in the given fasta file. It returns a MSA object from MIToS.MSA that keeps the original order of the sequences.

AlphaConformers.clean_msa — Method

clean_msa(msa::MIToS.MSA.AnnotatedMultipleSequenceAlignment)

This function removes the sequences with more than 50% of gaps and the columns with gaps in the reference sequence (the first one). The function returns a new MSA object.

AlphaConformers.create_msa_and_templates — Method

create_msa_and_templates(cluster_folder, ref_pdb, ref_chain, ref_model, 
    pdb_files, chains, models)

This function creates the folder structure defining AlphaFold 2 inputs, including the MSA and the templates. The cluster_folder argument is the path to the folder where the inputs will be created. The ref_pdb, ref_chain and ref_model arguments define the query structure. The pdb_files, chains and models arguments define the structures to use as templates. Note that the query structure or reference can also be included in pdb_files, chains and models; in that case, the ref_chain and ref_model arguments are ignored. The function returns the path to the pdb files, and their chains and models as well as the MSA. The returned values are a NamedTuple with the fields pdb_files, chains, models and msa.

AlphaConformers.create_pdb_folder — Method

create_pdb_folder(targets::Union{Set, DataFrames.DataFrame, AbstractArray}, 
                  folder_path::String; 
                  pdb_db::String = get(ENV, "PDB_DB", ""), 
                  extension::String = ".pdb")

Creates a local folder at folder_path and populates it with the PDB files specified by the targets. The targets argument can be a Set, an AbstractArray, or a DataFrame with a target column, such as the results loaded from a Foldseek search using the read_foldseek_search_results function. Each target represents a PDB code, possibly followed by a chain code. If a chain code is specified, only that chain is saved in the new folder. If no chain code is specified, all chains are saved. The expected format of the targets depends on the number of chains:

1 chain: "pdb"

2 chains: "pdb_chain"

For example: "1EX7.pdb" or "4F4J.pdb_B". Note that the presence of the "pdb" extension is not mandatory and depends on the database used when reading Foldseek results.

The PDB files are either fetched from a local database path specified by pdb_db or downloaded from the web if pdb_db is an empty string. Note that you can set the pdb_db keyword argument when calling the function, or alternatively, set the PDB_DB environment variable. When retrieving the PDB files from a local folder, the optional extension argument specifies the file extension of the PDB files in that folder (default is ".pdb").

Warning

If the specified folder_path already exists, it will be deleted and a new folder will be created.

The function returns the absolute path to the created folder.

AlphaConformers.create_pdb_lists — Method

create_pdb_lists(ref_pdb, ref_chain, ref_model, pdb_files, chains, models)

This function creates the lists of pdb files, chains and models to use as input for AlphaFold 2; defining the sequence and templates to consider.

AlphaConformers.foldseek_search — Method

foldseek_search(pdb_file::AbstractString; db_path::String = get(ENV, "FOLDSEEK_DB_PATH", ""), format_mode::Int = 0)

Searches a given protein structure in the Foldseek database.

The pdb_file is the path to the protein structure file in PDB format. The db_path is the path to the Foldseek database, which defaults to the FOLDSEEK_DB_PATH environment variable. If the FOLDSEEK_DB_PATH environment variable is not set and db_path is not given, the function will throw an error.

The function executes a Foldseek easy-search. It uses --format-mode 0 by default, which returns the results in the standard m8 format. Those results are saved in an output file with the same base name as the PDB file but with _results.m8 appended to it. The function returns the path to the Folseek easy-search output file.

If format_mode is set to 5, the function will return the path to the folder with the aligned structures. The aligned structures contains only the Calpha atoms aligned to the query structure.

The function operates in a temporary directory that is automatically cleaned up afterwards.

AlphaConformers.get_clustered_pdbs — Method

get_clustered_pdbs(clustering_result::StructuralClustering)

This function returns a list of lists of the PDB files in each cluster. The index of the outer list is the cluster number, and the inner list contains the PDB files.

AlphaConformers.get_pdb_codes — Method

get_pdb_codes(data::DataFrames.DataFrame, uni_acc::String)

This function returns a DataFrame with two columns, PDB and CHAIN, containing the PDB codes and chains associated with a given UniProt accession as uppercase strings. This function takes as input the DataFrame returned by the get_uniprot_mapping function as its first positional argument, data.

AlphaConformers.get_residues_and_sequence — Method

get_residues_and_sequence(pdb_file; chain=MIToS.PDB.All, model="1")

This function reads a pdb file and returns the residues and the sequence of the chain specified by the chain argument. The chain keyword argument can be a string with the chain name or the MIToS.PDB.All type. However, this function will throw an error if more than one chain is selected; so All is used only to avoid specifiying a chain when there is only one in the file. The model argument is used to select the model to read (by default the first model is selected).

AlphaConformers.get_uniprot_acc — Function

get_uniprot_acc(data::DataFrames.DataFrame, 
    pdb::String, chain::Union{String, Type{MIToS.PDB.All}}=MIToS.PDB.All)

This function returns a vector of UniProt accessions associated with a given PDB code and chain (if available). If the chain argument is not specified or if it is set to MIToS.PDB.All all UniProt accessions associated with the PDB code will be returned.

This function takes as input the DataFrame returned by the get_uniprot_mapping function as its first positional argument, data.

AlphaConformers.get_uniprot_mapping — Function

get_uniprot_mapping(sifts_db::String = get(ENV, "SIFTS_DB", pwd()))

Reads the pdb_chain_uniprot.csv.gz file and returns a DataFrame. If the file is not found in the specified sifts_db folder, it is downloaded from the SIFTS FTP server. When calling the function, you can set the sifts_db positional argument or the SIFTS_DB environment variable. If this is not done, the current working directory will be used by default.

AlphaConformers.read_foldseek_search_results — Method

read_foldseek_search_results(file::AbstractString)

Reads the Foldseek easy-search output file (m8) and returns a DataFrame with the results and proper column names.

AlphaConformers.run_alphafold — Method

run_alphafold(clusters_folder::String; colabfold_path::String=get(ENV, "COLABFOLD_PATH", ""))

This function runs AlphaFold 2 (colabfoldbatch) for each cluster in the `clustersfolder. Thecolabfoldpath` argument is the path to the colabfoldbatch program. Alternatively, you can set the COLABFOLD_PATH environment variable.

AlphaConformers.save_sequences — Method

save_sequences(cluster_folder::String, pdb_files::Vector{String}; chains, models)

This function uses the get_residues_and_sequence function to create a fasta file with the sequences of the structures in the cluster. The second arguments should be a vector with the paths to the pdb files of the structures in the cluster. The chains and models arguments are vectors with the chain and model to read for each pdb file. By default, the first model and all the chains are selected.

AlphaConformers.usalign — Method

usalign(pdb_file_a, pdb_file_b)
usalign(pdb_file_a, pdb_folder, pdb_list)

Performs structural alignment of protein structures using the USalign algorithm. The alignment is performed using the -fast option, which is slightly less accurate but faster than the default alignment method. The -TMcut 0.5 option is used to skip alignments unlikely to reach a TM-score of 0.5.

Function signatures

usalign(pdb_file_a, pdb_file_b) calculates the structural alignment between two protein structures. It takes the path to the two protein structures in PDB format as inputs.

usalign(pdb_file_a, pdb_folder, pdb_list) calculates the structural alignment between one protein structure and multiple other protein structures. It takes the path to the protein structure in PDB format to be aligned (pdb_file_a), the path to the directory containing the protein structures to align with (pdb_folder), and a list of protein structures (in PDB format) contained in pdb_folder (pdb_list) as inputs. You can use the create_pdb_folder function to create the pdb_folder using a list of target protein structures.

In both cases, the function asserts the existence of the input file(s) before proceeding with the alignment and returns a DataFrame with the results of the alignment.

USalign parameters

This function uses the USalign tool, interfaced via the USalign_jll package. The actual alignment is performed by running the tool in a subprocess, and the output is parsed from the tool's tabular output format (-outfmt 2). This function uses the following default parameters when running USalign:

-mol prot: Align proteins only

-TMcut 0.5: Skip alignment if TM-score is unlikely to reach 0.5 as we are looking for close conformations.

-fast: Use a faster but slightly less accurate alignment method

-outfmt 2: Use the tabular format for compact output that can be parsed with CSV.jl

Those parameters are also keyword arguments to the usalign function. You can use those keyword arguments to change this default behavior. To avoid using the TMscore cutoff, you should set TMcut to nothing. To use the slower but more accurate alignment method, you should set fast to false.

Also, you can pass a Cmd object to the extra_parameters keyword argument to add or modify other USalign arguments. Please note that you do not need to give the structures to align, as they are automatically passed to USalign.